应该完成了BPT，得到了正确的b

+++ code/components/metropolis_light_transport/include/PathContribution.hpp
@@ -4,7 +4,7 @@
 
 #include "Metropolis.hpp"
 
-#define PI ((float)3.1415926535898)
+#define PI ((double)3.1415926535898)
 #define MAX(x, y) ((x > y) ? x : y)
 #define MIN(x, y) ((x < y) ? x : y)
